\name{exact2x2-package}
\alias{exact2x2-package}
\docType{package}
\title{
Exact Tests and Confidence Intervals for 2x2 Tables
}
\description{
There are 4 main functions in the package. 
The \code{\link{exact2x2}} function calculates the exact conditional tests with matching confidence intervals as detailed in Fay (2010a,2010b). The functions \code{\link{ss2x2}} and \code{\link{power2x2}} calculate the sample size and power related to the tests of \code{exact2x2}. Finally, the \code{\link{binomMeld.test}} function calculates melded confidence intervals for two sample binomial inferences (see Fay, Proschan, and Brittain, 2014). 


 }
\details{
\tabular{ll}{
Package: \tab bpcp\cr
Type: \tab Package\cr
Version: \tab 1.3.0\cr
Date: \tab 2014-04-30\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Michael P. Fay

Maintainer: Michael P. Fay <mfay@niaid.nih.gov>

}
\references{
Fay, M. P. (2010a). Confidence intervals that Match Fisher's exact and Blaker's exact tests. Biostatistics,
  11: 373-374 (go to doc directory for earlier version or 
  \url{http://www3.niaid.nih.gov/about/organization/dcr/BRB/staff/michael.htm} for link to official version).

Fay, M.P. (2010b). Two-sided Exact Tests and Matching Confidence Intervals for Discrete Data. R Journal 2(1):53-58.


Fay, MP, Proschan, MA, and Brittain, E (2014). Combining One Sample Confidence Procedures for Inference in the Two Sample Case (unpublished manuscript). 
}
\keyword{ package }
\keyword{htest}
\keyword{nonparametric}
%\seealso{ }
%\examples{}
